+++ src/main/resources/com/dentalclinic/views/pages/manage/examination.fxml
@@ -9,80 +9,46 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<VBox alignment="CENTER" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" prefHeight="716.0" prefWidth="1054.0" stylesheets="@../../../css/style.css" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.dentalclinic.views.pages.manage.ExaminationPage">
-   <HBox prefHeight="2.0" prefWidth="1054.0" VBox.vgrow="ALWAYS">
+<VBox maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" prefHeight="716.0" prefWidth="1054.0" stylesheets="@../../../css/style.css" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.dentalclinic.views.pages.manage.ExaminationPage">
+   <Label prefHeight="57.0" prefWidth="214.0" styleClass="label-title" text="Examination">
+      <font>
+         <Font name="System Bold" size="18.0" />
+      </font>
+      <VBox.margin>
+         <Insets left="20.0" top="10.0" />
+      </VBox.margin>
+   </Label>
+   <HBox alignment="BOTTOM_RIGHT" prefHeight="52.0" prefWidth="217.0">
       <children>
-         <Label prefHeight="57.0" prefWidth="214.0" styleClass="label-title" text="Reg Number" HBox.hgrow="ALWAYS">
+         <Label fx:id="timeLabel" alignment="TOP_LEFT" prefHeight="34.0" prefWidth="263.0">
             <font>
-               <Font name="System Bold" size="18.0" />
+               <Font name="System Bold" size="14.0" />
             </font>
             <HBox.margin>
                <Insets top="10.0" />
             </HBox.margin>
          </Label>
-         <HBox alignment="BOTTOM_RIGHT" layoutX="159.0" layoutY="10.0" prefHeight="52.0" prefWidth="217.0" HBox.hgrow="ALWAYS">
-            <children>
-               <Label fx:id="timeLabel" alignment="TOP_LEFT" prefHeight="34.0" prefWidth="263.0">
-                  <font>
-                     <Font name="System Bold" size="14.0" />
-                  </font>
-                  <HBox.margin>
-                     <Insets top="10.0" />
-                  </HBox.margin>
-               </Label>
-               <TextField fx:id="searchField" prefHeight="38.0" prefWidth="160.0" promptText="Search" />
-            </children>
-            <padding>
-               <Insets right="30.0" />
-            </padding>
-         </HBox>
+         <TextField fx:id="searchField" prefHeight="38.0" prefWidth="160.0" promptText="Search" />
       </children>
-      <VBox.margin>
-         <Insets left="20.0" />
-      </VBox.margin>
+      <padding>
+         <Insets right="30.0" />
+      </padding>
    </HBox>
-   <VBox prefHeight="650.0" VBox.vgrow="ALWAYS">
-      <children>
-         <HBox layoutX="10.0" layoutY="50.0" prefHeight="491.0" prefWidth="1014.0" VBox.vgrow="ALWAYS">
-            <children>
-               <VBox HBox.hgrow="ALWAYS">
-                  <children>
-                     <VBox prefHeight="676.0" prefWidth="770.0">
-                        <children>
-                           <Label text="🏥 WAITING LIST OF PATIENTS">
-                              <font>
-                                 <Font name="Cambria Bold" size="14.0" />
-                              </font>
-                              <VBox.margin>
-                                 <Insets bottom="5.0" top="5.0" />
-                              </VBox.margin>
-                           </Label>
-                           <TableView fx:id="tableExamination" prefHeight="361.0" prefWidth="770.0" style="-fx-aligment: center;" VBox.vgrow="ALWAYS">
-                              <columns>
-                                 <TableColumn fx:id="nameColumn" prefWidth="137.60000228881836" text="Name" />
-                                 <TableColumn fx:id="reasonColumn" prefWidth="121.60002136230469" text="Reason" />
-                                 <TableColumn fx:id="symptomsColumn" minWidth="0.0" prefWidth="152.79998779296875" text="Symptoms" />
-                                 <TableColumn fx:id="doctorColumn" prefWidth="124.0" text="Doctor" />
-                                 <TableColumn fx:id="roomColumn" prefWidth="167.2000732421875" text="Room" />
-                                 <TableColumn fx:id="dateColumn" prefWidth="156.00006103515625" text="Date of visit" />
-                                 <TableColumn fx:id="actionColumn" prefWidth="139.19989013671875" text="Action" />
-                              </columns>
-                              <VBox.margin>
-                                 <Insets bottom="10.0" />
-                              </VBox.margin>
-                              <columnResizePolicy>
-                                 <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
-                              </columnResizePolicy>
-                           </TableView>
-                        </children>
-                     </VBox>
-                  </children>
-               </VBox>
-            </children>
-         </HBox>
-      </children>
+   <TableView fx:id="tableExamination" prefHeight="361.0" prefWidth="770.0" style="-fx-aligment: center;" VBox.vgrow="ALWAYS">
+      <columns>
+         <TableColumn fx:id="nameColumn" prefWidth="137.60000228881836" text="Name" />
+         <TableColumn fx:id="reasonColumn" prefWidth="121.60002136230469" text="Reason" />
+         <TableColumn fx:id="symptomsColumn" minWidth="0.0" prefWidth="152.79998779296875" text="Symptoms" />
+         <TableColumn fx:id="doctorColumn" prefWidth="124.0" text="Doctor" />
+         <TableColumn fx:id="roomColumn" prefWidth="167.2000732421875" text="Room" />
+         <TableColumn fx:id="dateColumn" prefWidth="156.00006103515625" text="Date of visit" />
+         <TableColumn fx:id="actionColumn" prefWidth="139.19989013671875" text="Action" />
+      </columns>
       <VBox.margin>
-         <Insets left="20.0" right="20.0" />
+         <Insets bottom="10.0" left="20.0" right="20.0" top="30.0" />
       </VBox.margin>
-   </VBox>
+      <columnResizePolicy>
+         <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
+      </columnResizePolicy>
+   </TableView>
 </VBox>
